Add searchNormalize option
Added 'searchNormalize' option to allow a user to have an option that normalizes a string by removing diacritic marks (accent marks) from the characters using the normalize() and replace() methods.

The normalize() method is called with the "NFD" argument, which decomposes the accented characters into a combination of a base character and a diacritic mark. For example, the letter "é" is decomposed into the base character "e" and the diacritic mark "`" (grave accent). This method works for most accented characters in most languages.
Associated with the issue with ID=238